pyproject.toml
[tool.poetry]
name = "krillboost"
version = "0.1.0"
description = ""
authors = ["Cian Kelly <cianrkelly23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.9,<3.12.4"
pandas = "^2.2.3"
numpy = "<2.0.0,>=1.23.0"
matplotlib = "^3.10.0"
xarray = "^2024.11.0"
copernicusmarine = "^1.3.5"
scikit-learn = "^1.6.0"
cmocean = "^4.0.3"
seaborn = "^0.13.2"
xgboost_cpu = "^2.1.4"

[tool.poetry.scripts]
exploreNC = "krillboost.preprocess.exploreNC:main"
downloadCop = "krillboost.preprocess.downloadCop:main"
subsetCop = "krillboost.preprocess.subsetCop:main"
fuseData = "krillboost.preprocess.dataFusion:main"
trainXG = "krillboost.training.train:main"
plotEnv = "krillboost.plotting.plotEnv:main"
plotClass = "krillboost.plotting.plotClass:main"
plotReg = "krillboost.plotting.plotReg:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
